+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # Changelog
+##0.5.2
+### Enhancements
+* injection works for `Manager`s
+* `initialize()` in the `Manager` is no longer abstract (same as in `EntitySystem`)
+* `World.createEntity` got an optional paramter to create an `Entity` with components
+* new function `World.createAndAddEntity` which adds the `Entity` to the world after creation
+
+### Bugfixes
+* added getter for the `World` in `Manager` 
+* the uniqueId of an `Entity` was always 0, not very unique
+
 ##0.5.1
 ### Internal
 * added version constraint for release of Dart